requirements.txt
pydantic-ai
tavily-python
streamlit
python-dotenv
openai
logfire